requirements.txt
django==2.2.10
djangorestframework==3.11.0
drf-yasg==1.17.1